Added graphics to level 1, player and enemy

+++ src/org/academiadecodigo/secondrow/killbox/objects/enemy/LaserEnemy.java
@@ -1,7 +1,6 @@
 package org.academiadecodigo.secondrow.killbox.objects.enemy;
 
 import org.academiadecodigo.secondrow.graphics.*;
-import org.academiadecodigo.secondrow.killbox.Var;
 import org.academiadecodigo.secondrow.killbox.maps.Map;
 import org.academiadecodigo.secondrow.killbox.objects.Collidable;
 import org.academiadecodigo.secondrow.killbox.objects.Position;
@@ -68,38 +67,19 @@ public class LaserEnemy extends Enemy implements Collidable {
 
     }
 
-    @Override
-    public void delete(){
-
-        enemy.delete();
-        shot.delete();
-
-    }
 
 
 
     public void shot(int playerX, int playerY){
 
 
-        //changed from 900 to 300
-        if(shotTaken && counter < 300) {
+
+        if(shotTaken && counter < 900) {
             counter++;
             return;
         }
-
-        if(counter == 300){shot.delete();}
-
-        int inicialX = pos.getX() + (width/2);
-        int inicialY = pos.getY() + (height/2);
-
-        int finalY = Var.HEIGHT - Var.WALL_PADDING + 10;
-
-        int finalX = playerX + Var.PLAYER_WIDTH;
-
-
-        //ADD line-platform collision and ADD player-line collision and ADD don't shot when below platform
-
-        shot = new Line( inicialX, inicialY, finalX, finalY);
+        if(counter == 900){shot.delete();}
+        shot = new Line(pos.getX() + (width/2), pos.getY() + (height/2), playerX, playerY);
         shot.setColor(Color.GREEN);
         shot.draw();
         shotTaken = true;
